refactored image view animation code into its own extension file added comments for animation code
